move fireball listeners to a seperate class
