+++ bedwars-plugin/src/main/java/com/andrei1058/bedwars/configuration/MainConfig.java
@@ -101,7 +101,6 @@ public class MainConfig extends ConfigManager {
         yml.addDefault(ConfigPath.GENERAL_TNT_JUMP_DAMAGE_TEAMMATES, 5);
         yml.addDefault(ConfigPath.GENERAL_TNT_JUMP_DAMAGE_OTHERS, 10);
         // fireball category
-        yml.addDefault(ConfigPath.GENERAL_FIREBALL_DAMAGE_MULTIPLIER, 0.21);
         yml.addDefault(ConfigPath.GENERAL_FIREBALL_EXPLOSION_SIZE, 3);
         yml.addDefault(ConfigPath.GENERAL_FIREBALL_SPEED_MULTIPLIER, 10);
         yml.addDefault(ConfigPath.GENERAL_FIREBALL_MAKE_FIRE, false);
@@ -211,6 +210,10 @@ public class MainConfig extends ConfigManager {
                     set(ConfigPath.GENERAL_CONFIGURATION_ARENA_SELECTOR_STATUS_ENCHANTED.replace("%path%", new_path), getYml().getBoolean("arenaGui." + path + ".enchanted"));
                 }
             }
+        }
+
+        if(getYml().get("fireball.damage-multiplier") != null) {
+            set("fireball.damage-multiplier", null);
         }
 
         set("arenaGui", null);
