+++ bedwars-plugin/src/main/java/com/andrei1058/bedwars/listeners/DamageDeathMove.java
@@ -70,12 +70,6 @@ public class DamageDeathMove implements Listener {
     private final double tntDamageTeammates;
     private final double tntDamageOthers;
 
-    private final double fireballDamageMultiplier;
-    private final double fireballExplosionSize;
-    private final boolean fireballMakeFire;
-    private final double fireballHorizontal;
-    private final double fireballVertical;
-
     public DamageDeathMove() {
         this.tntJumpBarycenterAlterationInY = config.getYml().getDouble(ConfigPath.GENERAL_TNT_JUMP_BARYCENTER_IN_Y);
         this.tntJumpStrengthReductionConstant = config.getYml().getDouble(ConfigPath.GENERAL_TNT_JUMP_STRENGTH_REDUCTION);
@@ -83,12 +77,6 @@ public class DamageDeathMove implements Listener {
         this.tntDamageSelf = config.getYml().getDouble(ConfigPath.GENERAL_TNT_JUMP_DAMAGE_SELF);
         this.tntDamageTeammates = config.getYml().getDouble(ConfigPath.GENERAL_TNT_JUMP_DAMAGE_TEAMMATES);
         this.tntDamageOthers = config.getYml().getDouble(ConfigPath.GENERAL_TNT_JUMP_DAMAGE_OTHERS);
-
-        this.fireballDamageMultiplier = config.getYml().getDouble(ConfigPath.GENERAL_FIREBALL_DAMAGE_MULTIPLIER);
-        this.fireballExplosionSize = config.getYml().getDouble(ConfigPath.GENERAL_FIREBALL_EXPLOSION_SIZE);
-        this.fireballMakeFire = config.getYml().getBoolean(ConfigPath.GENERAL_FIREBALL_MAKE_FIRE);
-        this.fireballHorizontal = config.getYml().getDouble(ConfigPath.GENERAL_FIREBALL_KNOCKBACK_HORIZONTAL);
-        this.fireballVertical = config.getYml().getDouble(ConfigPath.GENERAL_FIREBALL_KNOCKBACK_VERTICAL);
     }
 
     @EventHandler
@@ -123,10 +111,6 @@ public class DamageDeathMove implements Listener {
                 }
                 //}
 
-                // Fireball is BLOCK_EXPLOSION because of spawned explosion
-                if(e.getCause() == EntityDamageEvent.DamageCause.BLOCK_EXPLOSION) {
-                    e.setDamage(e.getDamage()*fireballDamageMultiplier);
-                }
             }
         }
         if (BedWars.getServerType() == ServerType.MULTIARENA) {
@@ -134,67 +118,6 @@ public class DamageDeathMove implements Listener {
                 e.setCancelled(true);
             }
         }
-    }
-
-    Map<Location, Player> explosionSources = new HashMap<>();
-
-    @EventHandler
-    public void fireballPrime(ExplosionPrimeEvent e) {
-        if(!(e.getEntity() instanceof Fireball)) return;
-        ProjectileSource shooter = ((Fireball) e.getEntity()).getShooter();
-        if(!(shooter instanceof Player)) return;
-        IArena a = Arena.getArenaByPlayer((Player) shooter);
-        if(a == null) return;
-        e.setRadius((float) fireballExplosionSize);
-        e.setFire(fireballMakeFire);
-        explosionSources.put(e.getEntity().getLocation(), (Player) shooter);
-    }
-
-    @EventHandler
-    public void fireballExplode(EntityExplodeEvent e) {
-        if(!(e.getEntity() instanceof Fireball)) return;
-        Location location = e.getLocation();
-
-        Player source = null;
-        for(Location sourceLocation : explosionSources.keySet()) {
-            if(sourceLocation == null) continue;
-            if(sourceLocation.distance(location) < 0.05) {
-                source = explosionSources.get(sourceLocation);
-                explosionSources.remove(sourceLocation);
-            }
-        }
-
-        if(source == null) return;
-
-        Vector vector = location.toVector();
-
-        Collection<Entity> nearbyEntities = location.getWorld()
-                .getNearbyEntities(location, fireballExplosionSize, fireballExplosionSize, fireballExplosionSize);
-        for(Entity entity : nearbyEntities) {
-            if(!(entity instanceof Player)) continue;
-            Player player = (Player) entity;
-            if(!getAPI().getArenaUtil().isPlaying(player)) continue;
-            IArena arena = Arena.getArenaByPlayer(player);
-
-
-            Vector playerVector = player.getLocation().toVector();
-            Vector normalizedVector = vector.subtract(playerVector).normalize();
-            Vector horizontalVector = normalizedVector.multiply(fireballHorizontal);
-            double y = normalizedVector.getY();
-            if(y < 0 ) y += 1.5;
-            if(y <= 0.5) {
-                y = fireballVertical*1.5; // kb for not jumping
-            } else {
-                y = y*fireballVertical*1.5; // kb for jumping
-            }
-            player.setVelocity(horizontalVector.setY(y));
-
-            // dont damage teammates
-            if(!player.equals(source) && arena.getTeam(player).equals(arena.getTeam(source))) continue;
-
-            player.damage(2.0);
-        }
-
     }
 
     // show player health on bow hit
